# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs API
  contact: {}
  version: '1.0'
servers:
- url: https://jobify-app-api-dev.onrender.com/api/v1
  variables: {}
paths:
...
  /auth/user/{email}:
    parameters:
      - in: path
        name: email
        schema:
          type: string
        required: true
        description: The user email
    delete:
      tags:
      - Auth
      summary: Delete User
      operationId: DeleteUser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...